requirements-minimal.txt
# Minimal Requirements - Nur für Grundfunktionen
# Diese Pakete kompilieren ohne C-Compiler

# Datenbank
mysql-connector-python==8.3.0

# FinTS / Banking
fints==4.0.0
mt-940==4.30.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Utils
colorlog==6.8.0
python-dateutil==2.8.2

# Optional (später mit C-Compiler):
# pandas==2.2.0
# numpy==1.26.3
# PyPDF2==3.0.1
# pdfplumber==0.10.3
# pytesseract==0.3.10
# sqlalchemy==2.0.25
# scikit-learn==1.4.0